钉钉工作通知参数和响应结果

+++ tripartite-ding/tripartite-ding-openapi/src/main/java/work/gaigeshen/tripartite/ding/openapi/parameters/process/DingProcessWorkRecordTaskUpdateParameters.java
@@ -28,7 +28,7 @@ public class DingProcessWorkRecordTaskUpdateParameters implements ClientParamete
 
     public static class Task {
 
-        public Integer task_id;
+        public Long task_id;
 
         public String status;
