+++ tripartite-ding/tripartite-ding-openapi/src/main/java/work/gaigeshen/tripartite/ding/openapi/response/process/DingProcessWorkRecordTaskCreateResponse.java
@@ -13,7 +13,7 @@ public class DingProcessWorkRecordTaskCreateResponse implements ClientResponse {
 
     public static class Task {
 
-        public Integer task_id;
+        public Long task_id;
 
         public String userid;
     }
